+++ l10n_ir_base/models/__init__.py
@@ -1,4 +1,6 @@
 # from . import sale_report
+from . import res_user
+from . import res_lang
 from . import base_model
+from . import res_partner
 from . import account_journal_dashboard
-from . import user_extensions
