+++ l10n_ir_base/models/res_lang.py
@@ -1,0 +1,13 @@
+from odoo import fields, models
+
+
+class UsersExtensions(models.Model):
+    _inherit = "res.lang"
+
+
+    calendar = fields.Selection(
+        [("gregorian", "Gregorian Calendar"), ("jalali", "Jalali (Persian) Calendar")],
+        default="jalali",
+    )
+
+
