+++ l10n_ir_base/views/res_lang_form.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <data>
+        <record model="ir.ui.view" id="l10n_ir_res_lang_form">
+          <field name="name">Lang Calendar</field>
+          <field name="model">res.lang</field>
+          <field name="inherit_id" ref="base.res_lang_form" />
+          <field name="arch" type="xml">
+            <field name="date_format" position="before">
+              <field name='calendar' string='Calendar' />
+            </field>
+          </field>
+        </record>
+    </data>
+</odoo>
