+++ l10n_ir_base/views/res_partner_form.xml
@@ -19,3 +19,9 @@
             </field>
         </record>
 </odoo>
+
+
+
+
+
+
